Transfer filters to util package
