+++ src/main/java/school/faang/user_service/service/RecommendationRequestService.java
@@ -12,7 +12,7 @@ import school.faang.user_service.mapper.recommendation.RecommendationRequestMapp
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.recommendation.RecommendationRequestRepository;
 import school.faang.user_service.repository.recommendation.SkillRequestRepository;
-import school.faang.user_service.service.reccomendation.filter.RecommendationRequestFilter;
+import school.faang.user_service.util.filter.recommendationRequest.RecommendationRequestFilter;
 
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
