+++ src/main/java/school/faang/user_service/util/filter/recommendationRequest/RecommendationRequestFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.reccomendation.filter;
+package school.faang.user_service.util.filter.recommendationRequest;
 
 import school.faang.user_service.dto.recommendation.filter.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
