+++ src/main/java/school/faang/user_service/util/filter/recommendationRequest/RecommendationRequestMessageFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.service.reccomendation.filter;
+package school.faang.user_service.util.filter.recommendationRequest;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.recommendation.filter.RequestFilterDto;
@@ -7,18 +7,15 @@ import school.faang.user_service.entity.recommendation.RecommendationRequest;
 import java.util.stream.Stream;
 
 @Component
-public class RecommendationRequestCreatedAtFilter implements RecommendationRequestFilter {
+public class RecommendationRequestMessageFilter implements RecommendationRequestFilter {
     @Override
     public boolean isApplicable(RequestFilterDto filters) {
-        return filters.getCreatedAt() != null;
+        return filters.getMessage() != null && !filters.getMessage().isBlank();
     }
 
     @Override
     public Stream<RecommendationRequest> apply(Stream<RecommendationRequest> requests, RequestFilterDto filters) {
         return requests
-                .filter(request -> request
-                        .getCreatedAt()
-                        .toLocalDate()
-                        .isEqual(filters.getCreatedAt().toLocalDate()));
+                .filter(request -> request.getMessage().contains(filters.getMessage().trim()));
     }
 }
